refactor: lombok 적용
